+++ packages/cli/src/cli/commands/upload.ts
@@ -10,14 +10,14 @@ import { logger } from '../../console/logger.js';
 import { getRelative, readFile } from '../../fs/findFilepath.js';
 import { ResolvedFiles, Settings, TransformFiles } from '../../types/index.js';
 import { FileFormat, DataFormat } from '../../types/data.js';
-import { SUPPORTED_FILE_EXTENSIONS } from './supportedFiles.js';
-import { UploadOptions } from '../../cli/base.js';
+import { SUPPORTED_FILE_EXTENSIONS } from '../../formats/files/supportedFiles.js';
+import { UploadOptions } from '../base.js';
 import sanitizeFileContent from '../../utils/sanitizeFileContent.js';
-import { parseJson } from '../json/parseJson.js';
+import { parseJson } from '../../formats/json/parseJson.js';
 import { runUploadFilesWorkflow } from '../../workflows/upload.js';
 import { existsSync, readFileSync } from 'node:fs';
-import { createFileMapping } from './fileMapping.js';
-import parseYaml from '../yaml/parseYaml.js';
+import { createFileMapping } from '../../formats/files/fileMapping.js';
+import parseYaml from '../../formats/yaml/parseYaml.js';
 import type { FileToUpload } from 'generaltranslation/types';
 import { hashStringSync } from '../../utils/hash.js';
