+++ packages/cli/src/formats/files/aggregateFiles.ts
